Typecasted the value inside floor to float

+++ core/math/math_funcs.cpp
@@ -119,7 +119,7 @@ double Math::ease(double p_x, double p_c) {
 
 double Math::snapped(double p_value, double p_step) {
 	if (p_step != 0) {
-		p_value = Math::floor(p_value / p_step + 0.5) * p_step;
+		p_value = Math::floor((float)(p_value / p_step + 0.5)) * p_step;
 	}
 	return p_value;
 }
